Added refinement to all C libs.
Note that the c implementations are righ tnow working but are not
very effective, next step is to make them much faster!

+++ src/c_lib/optimize/solvers.h
@@ -3,14 +3,14 @@
 #include <fftw3.h>
 
 
-float complex dtft_fractor_sum(float freq, const fftwf_complex* decoded_signal, float sample_rate, int signal_len);
-float dtft_optimize_fun(float freq, const fftwf_complex* decoded_signal, float sample_rate, int signal_len);
-float dtft_solve(
+double complex dtft_fractor_sum(double freq, const fftwf_complex* decoded_signal, float sample_rate, int signal_len);
+double dtft_optimize_fun(double freq, const fftwf_complex* decoded_signal, float sample_rate, int signal_len);
+double dtft_solve(
     const fftwf_complex* decoded_signal,
     int signal_len,
     float sample_rate,
-    float freq_start,
-    float freq_end,
-    float* fmin_pwr,
+    double freq_start,
+    double freq_end,
+    double* fmin_pwr,
     double* phi
 ) ;
